chef/backups/backup100725/chefmain/requirements.txt
certifi==2024.12.14
colorama==0.4.6
pinecone==5.4.2
pinecone-plugin-inference==3.1.0
pinecone-plugin-interface==0.0.7
python-dateutil==2.9.0.post0
six==1.17.0
tqdm==4.67.1
typing-extensions==4.12.2
urllib3==2.2.3
requests==2.32.3
flask==3.0.3
numpy==2.1.1
openai==1.45.1
replit==4.1.0
gspread==6.1.2
aiohttp==3.10.5
cryptography==43.0.1
firebase-admin==6.4.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
pytz==2024.1
python-telegram-bot==20.7
serpapi==0.1.5
google-search-results==2.4.2
watchdog==3.0.0
nest_asyncio
psutil==5.9.5
python_dotenv==1.0.1
python-telegram-bot[webhooks]
